backend/django/requirements.txt
Django==4.2
djangorestframework==3.14.0
psycopg==3.1.8
django-cors-headers==3.14.0
djangorestframework-simplejwt==5.2.2
pytest==7.2.2
pytest-django==4.5.2
dj-database-url==2.1.0
weasyprint==52.5
channels==4.3.1
channels-redis==4.2.0
django-otp==1.6.1
qrcode[pil]==8.2
pyotp==2.9.0
daphne==4.2.1
python-socketio==5.14.1
python-engineio==4.12.3
eventlet==0.40.3
requests==2.32.5
stripe==13.0.1
paypalrestsdk==1.13.3